Changed project layout
Moved the `main.py` file and the pages directory from the
`streamlit_passwordless_app` package to the root level of the
project directory. This should solve the issue of deploying the
application on Streamlit Community Cloud, where you need to have
the entry point Python file at the root level of the project. The
pages folder have to be on the same level as the entry point file.