Updated view register_and_sign_in_section
Replaced the sign in form with the newly added `streamlit_passwordless.sign_in_button`.
Merged the register and sign in tabs and placed the sign in button below
the register form. This looks more elegant. Removed the possibility to
toggle between discoverable and non-discoverable passkeys since the normal
user will rarely ever need to create a non-discoverable passkey and the
difference between them can be confusing to new users.